feat: extract shared planner sidebar logic
